Remove default allocations for Vector3 `dest` params
Previously, all `dest` parameters in Vector3 functions were
optional, and new Vector3s would be allocated if the destination
was not provided. As of this commit, `dest` is no longer optional.
This was done in the interest of removing overlooked allocations,
and also in cutting down on the overhead needed to check if the
`dest` parameter is undefined.